Fix vignette tooltip rewards [#89]
Vignette tooltips in the game no longer call the SetText function on
the GameTooltip. They instead call AddLine, which is not as safe to
hook because AddLine can be called multiple times before displaying
the tooltip, which would cause the rewards to get added multiple times.

Instead, we now hook the Show function to check if rewards need to be
added after the finished tooltip is shown.