Add rlabel icon for rift rares
Also include notes about which rares pop out and which rares
stay in the rift after pulling them. Since 3/4 rares in Korthia
pop out (when they fucking work Blizzard), it now makes more sense
to display them in both phases rather than leave Yorik as the only
rare in Korthia that is only seen when in the rift phase.