Add icon for tormentors of torghast
I chose not to include anything about the schedule or locations. I'm
not confident I can get the date math exactly right 100% of the time,
and if Blizzard changes anything we're then committed to figuring out
the new schedule. Instead we'll just the three key pieces of information
people care about in the one giant tooltip:

  1. Have I looted the weekly box for cinders?
  2. What is the order of the bosses so I know what is spawning next?
  3. What are the rewards from each boss and the weekly box?