Add option to show unobtainable transmogs
Transmogs that are not obtainable by your current character are
now hidden by default. An option called 'Show unobtainable transmogs'
has been added to the global options to restore the previous tooltip
behavior.

The giant tormentor and mirror tooltips have increased the need for
this feature, since its not really useful to have a tooltip filled
with 75%+ transmogs that your character does not care about.