Create LICENSE
Not picking a generic open source license given the nature of this project and the recent problems we've been facing. This file makes explicit the implicit "All rights reserved" of any project without a license defined.

I may open the project up with an open source license in the future if/when I decide I no longer want to be its maintainer, in order to let others continue the work.