Update enUS.lua
Blizzard made the quest(s) required to acquire the temporary Honorary Dryad title repeatable, so anyone losing the title before they can claim the pet are now able to reacquire the title again at a later point. I'm not sure if they're daily or weekly repeatable, but they are repeatable now.

I changed the bug-note to a note highlighting the fact that you can now repeat it. Change it to whatever you want if my edit isn't suitable.