Cosmetic update (#408)
* Update for cosmetic transmog

* Add tabard localization
* Check if an item is cosmetic and if it is, it will now be displayed in the tooltip + real slot
* Update many transmog rewards

* Add isCosmetic for Transmog

Only a Transmog can be cosmetic

* Add Forbidden Hoard vignette ids

* Fix for multiple rendered vignette rewards

This fixes that rewards are rendered multiple times in the Gametooltip if multiple nodes have the same vignette id e. g. Forbidden Hoard.

* Update Transmog (double check)

* Change Transmog type to slot

we use slot for Transmog in all plugins, but these used type

* Check in Transmog:IsObtainable() for isCosmetic

This must have been a problem when we had slot = L['cloak'], but the item was a cosmetic.

I checked this on torghast cosmetics.